Common functions

UHD.openUHD — Function

Init the core parameter of the radio in Tx or in Rx mode and initiate RF parameters

–- Syntax

open(mode,sysImage,carrierFreq,samplingRate,txGain,antenna="RX2")

–- Input parameters

mode : String to open radio in "Tx" (transmitter) or in "Rx" (receive) mode

sysImage : String with the additionnal load parameters (for instance, path to the FPHGA image) [String]

carrierFreq : Desired Carrier frequency [Union{Int,Float64}]

samplingRate : Desired bandwidth [Union{Int,Float64}]

txGain : Desired Tx Gain [Union{Int,Float64}]

antenna : Desired Antenna alias [String]

–- Output parameters

UHDTx : UHD Tx or Rx object with PHY parameters [Union{UHDTx,UHDRx}]

Receiver functions

Base.close — Method

Close the USRP device (Rx mode) and release all associated objects

–- Syntax

close(uhd)

–- Input parameters

uhd : UHD object [UHDRx]

–- Output parameters

[]

Base.print — Method

Print the radio configuration

–- Syntax

printUHD(radio)

–- Input parameters

radio : UHD object (Tx or Rx)

–- Output parameters

[]

UHD.getError — Method

Returns the Error flag of the current UHD burst

–- Syntax

flag = getError(radio)

–- Input parameters

radio : UHD object [UHDRx]

–- Output parameters

err : Error Flag [errorcodet]

UHD.getTimestamp — Method

Return the timestamp of the last UHD burst

–- Syntax

(second,fracSecond) = getTimestamp(radio)

–- Input parameters

radio : UHD UHD object [UHDRx]

–- Output parameters

second : Second value for the flag [Int]

fracSecond : Fractional second value [Float64]

UHD.initRxUHD — Method

Initiate all structures to instantiate and pilot a USRP device into Receiver mode (Rx).

–- Syntax

uhd = initRxUHD(sysImage)

–- Input parameters

sysImage : String with the additionnal load parameters (for instance, path to the FPHGA image) [String]

–- Output parameters

uhd = UHD Rx object [UHDRxWrapper]

UHD.openUHDRx — Function

Init the core parameter of the radio (Rx mode) and initiate RF parameters

–- Syntax

openUHDRx(sysImage,carrierFreq,samplingRate,gain,antenna="TX-RX")

–- Input parameters

sysImage : String with the additionnal load parameters (for instance, path to the FPHGA image) [String]

carrierFreq : Desired Carrier frequency [Union{Int,Float64}]

samplingRate : Desired bandwidth [Union{Int,Float64}]

gain : Desired Rx Gain [Union{Int,Float64}]

antenna : Desired Antenna alias (default "TX-RX") [String]

–- Output parameters

UHDRx : UHD Rx object with PHY parameters [UHDRx]

UHD.populateBuffer! — Function

Calling UHD function wrapper to fill a buffer

–- Syntax

recv!(sig,radio,nbSamples)

–- Input parameters

radio : UHD object [UHDRx]

ptr : Writable memory position [Ref{Ptr{Cvoid}}]

nbSamples : Number of samples to acquire

–- Output parameters

nbSamp : Number of samples fill in buffer [Csize_t]

UHD.recv! — Method

Get a single buffer from the USRP device, using the Buffer structure

–- Syntax

recv!(sig,radio,nbSamples)

–- Input parameters

sig : Complex signal to populate [Array{Complex{Cfloat}}]

radio : UHD object [UHDRx]

buffer : Buffer object (obtained with setBuffer(radio)) [Buffer]

–- Output parameters

sig : baseband signal from radio [Array{Complex{Cfloat}},radio.packetSize]

UHD.recv — Method

Get a single buffer from the USRP device, and create all the necessary ressources

–- Syntax

sig = recv(radio,nbSamples)

–- Input parameters

radio : UHD object [UHDRx]

nbSamples : Desired number of samples [Int]

–- Output parameters

sig : baseband signal from radio [Array{Complex{CFloat}},radio.packetSize]

UHD.updateCarrierFreq! — Method

Update carrier frequency of current radio device, and update radio object with the new obtained carrier frequency

–- Syntax

updateCarrierFreq!(radio,carrierFreq)

–- Input parameters

radio : UHD device [UHDRx]

carrierFreq : New desired carrier freq

–- Output parameters

carrierFreq : Current radio carrier frequency

–-

v 1.0

UHD.updateGain! — Method

Update gain of current radio device, and update radio object with the new obtained gain

–- Syntax

updateGain!(radio,gain)

–- Input parameters

radio : UHD device [UHDRx]

gain : New desired gain

–- Output parameters

gain : Current radio gain

UHD.updateSamplingRate! — Method

Update sampling rate of current radio device, and update radio object with the new obtained sampling frequency

–- Syntax

updateSamplingRate!(radio,samplingRate)

–- Input parameters

radio : UHD device [UHDRx]

samplingRate : New desired sampling rate

–- Output parameters

Transmitter functions

Base.close — Method

Close the USRP device (Tx mode) and release all associated objects

–- Syntax

close(uhd)

–- Input parameters

uhd : UHD object [UHDTx]

–- Output parameters

[]

UHD.initTxUHD — Method

Initiate all structures to instantiate and pilot a USRP device in Transmitter (Tx) mode.

–- Syntax

uhd = initTxUHD(sysImage)

–- Input parameters

sysImage : String with the additionnal load parameters (for instance, path to the FPHGA image) [String]

–- Output parameters

uhd = UHD Tx object [UHDTxWrapper]

UHD.openUHDTx — Function

Init the core parameter of the radio in Tx mode and initiate RF parameters

–- Syntax

openUHDTx(sysImage,carrierFreq,samplingRate,gain,antenna="TX-RX")

–- Input parameters

sysImage : String with the additionnal load parameters (for instance, path to the FPHGA image) [String]

carrierFreq : Desired Carrier frequency [Union{Int,Float64}]

samplingRate : Desired bandwidth [Union{Int,Float64}]

gain : Desired Tx Gain [Union{Int,Float64}]

antenna : Desired Antenna alias (default "TX-RX") [String]

–- Output parameters

UHDTx : UHD Tx object with PHY parameters [UHDTx]

UHD.send — Function

Send a buffer though the radio device. It is possible to force a cyclic buffer send (the radio uninterruply send the same buffer) by setting the cyclic parameter to true

–- Syntax

send(radio,buffer,cyclic=false)

–- Input parameters

radio : UHD device [UHDRx]

buffer : Buffer to be send [Union{Array{Complex{Cfloat}},Array{Cfloat}}]

cyclic : Send same buffer multiple times (default false) [Bool]

–- Output parameters

nbEch : Number of samples effectively send [Csize_t]. It corresponds to the number of complex samples sent.

–-

v 1.0

UHD.updateCarrierFreq! — Method

Update carrier frequency of current radio device, and update radio object with the new obtained carrier frequency

–- Syntax

updateCarrierFreq!(radio,carrierFreq)

–- Input parameters

radio : UHD device [UHDRx]

carrierFreq : New desired carrier freq

–- Output parameters

carrierFreq : Current radio carrier frequency

–-

v 1.0

UHD.updateGain! — Method

Update gain of current radio device, and update radio object with the new obtained gain

–- Syntax

updateGain!(radio,gain)

–- Input parameters

radio : UHD device [UHDTx]

gain : New desired gain

–- Output parameters

updateGain : Current Radio gain [Float64]

UHD.updateSamplingRate! — Method

Update sampling rate of current radio device, and update radio object with the new obtained sampling frequency

–- Syntax

updateSamplingRate!(radio,samplingRate)

–- Input parameters

radio : UHD device [UHDTx]

samplingRate : New desired sampling rate

–- Output parameters